RD-54 integrate paypal

+++ gqschema/coupon.js
@@ -5,7 +5,7 @@ module.exports = gql`
     code: String
     description: String
     discount_type: String
-    discount_value: Int
+    discount_value: Float
     free_shipping: Boolean
     expire: String
     minimum_spend: Int
@@ -51,7 +51,7 @@ module.exports = gql`
       code: String
       description: String
       discount_type: String
-      discount_value: Int
+      discount_value: Float
       free_shipping: Boolean
       expire: String
       minimum_spend: Int
@@ -68,7 +68,7 @@ module.exports = gql`
       code: String
       description: String
       discount_type: String
-      discount_value: Int
+      discount_value: Float
       free_shipping: Boolean
       expire: String
       minimum_spend: Int
